LexiScan_Auto/requirements.txt
pytesseract
pdf2image
pillow
transformers
torch
spacy
pandas
numpy
scikit-learn
fastapi
uvicorn